server/requirements.txt
Flask
flask-cors
flask-jwt-extended
requests
sse-starlette
psycopg2-binary
SQLAlchemy
pgvector
pdfminer.six
PyMuPDF
camelot-py[cv]
tabula-py
paddlepaddle
paddleocr
pytesseract
youtube-transcript-api
yt-dlp
faster-whisper
vosk
pyttsx3
langdetect
